+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.26.0 (2024-08-28)
+
+### Added
+
+- Update `SideBar` visual
+
+## 1.25.0 (2024-08-16)
+
+### Added
+
+- `DropArea.vue` component from `UploadArea` component.
+
+### Changed
+
+- Update `ImportCard.vue` design.
+
 ## 1.24.22 (2024-06-20)
 
 ### Fixed